+++ CriticalMass/RequestManager.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import os.log
 
 public class RequestManager {
     private struct SendLocationPostBody: Codable {
@@ -27,6 +28,8 @@ public class RequestManager {
     private var networkLayer: NetworkLayer
     private var idProvider: IDProvider
 
+    private var log = OSLog(subsystem: Bundle.main.bundleIdentifier!, category: "RequestManager")
+
     init(dataStore: DataStore, locationProvider: LocationProvider, networkLayer: NetworkLayer, interval: TimeInterval = 12.0, idProvider: IDProvider, url: URL) {
         endpoint = url
         self.idProvider = idProvider
@@ -41,20 +44,30 @@ public class RequestManager {
     }
 
     @objc private func timerDidUpdate(timer _: Timer) {
+        Logger.log(.info, log: log, "Timer did update")
         updateData()
     }
 
     private func defaultCompletion(for response: ApiResponse?) {
-        if let response = response {
-            DispatchQueue.main.async {
+        DispatchQueue.main.async {
+            defer {
                 self.hasActiveRequest = false
+            }
+            if let response = response {
                 self.dataStore.update(with: response)
+
+                Logger.log(.info, log: self.log, "Successfully finished API update")
+            } else {
+                Logger.log(.error, log: self.log, "API update failed")
             }
         }
     }
 
     private func updateData() {
-        guard hasActiveRequest == false else { return }
+        guard hasActiveRequest == false else {
+            Logger.log(.info, log: log, "Don't attempt to request new data because because a request is still active")
+            return
+        }
         hasActiveRequest = true
         // We only use a post request if we have a location to post
         if let currentLocation = locationProvider.currentLocation {